+++ samples/sampler/stories/qa/Scroller.js
@@ -242,6 +242,27 @@ storiesOf('Scroller', module)
 		)
 	)
 	.add(
+		'With short contents',
+		() => (
+			<Scroller
+				direction={select('direction', prop.direction, Config)}
+				focusableScrollbar={prop.focusableScrollbarOption[select('focusableScrollbar', ['false', 'true', '"byEnter"'], Config)]}
+				horizontalScrollbar={select('horizontalScrollbar', prop.scrollbarOption, Config)}
+				key={select('scrollMode', prop.scrollModeOption, Config)}
+				noScrollByWheel={boolean('noScrollByWheel', Config)}
+				onKeyDown={action('onKeyDown')}
+				onScrollStart={action('onScrollStart')}
+				onScrollStop={action('onScrollStop')}
+				scrollMode={select('scrollMode', prop.scrollModeOption, Config)}
+				spotlightDisabled={boolean('spotlightDisabled', Config, false)}
+				style={{height: ri.scaleToRem(600)}}
+				verticalScrollbar={select('verticalScrollbar', prop.scrollbarOption, Config)}
+			>
+				Text
+			</Scroller>
+		)
+	)
+	.add(
 		'With Resizable',
 		() => (
 			<ScrollerWithResizable />